update
I found out the reason why the stylesheet wasn't working it was because of the link, I didnt put it correctly but now it works perfectly, another thing that I had issues was the Heroku not working, the reason was that I had multiple domains and github didn't know what to do so I had to delete all of them before the app to work.  I added licenses in the app, the description with screenshots and videos of the app, and the LICENSE file to give credit to those that help me achieve the app.